Rename addon vitest to addon test
