+++ code/addons/test/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@storybook/experimental-addon-vitest",
+  "name": "@storybook/experimental-addon-test",
   "version": "8.3.0-beta.0",
   "description": "Integrate Vitest with Storybook",
   "keywords": [
@@ -29,7 +29,7 @@
       "import": "./dist/index.js",
       "require": "./dist/index.cjs"
     },
-    "./plugin": {
+    "./vite-plugin": {
       "types": "./dist/plugin/index.d.ts",
       "import": "./dist/plugin/index.js",
       "require": "./dist/plugin/index.cjs"
