+++ code/addons/test/project.json
@@ -1,5 +1,5 @@
 {
-  "name": "vitest",
+  "name": "addon-test",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
   "projectType": "library",
   "targets": {
