+++ code/addons/test/src/manager.tsx
