+++ code/addons/test/src/plugin/global-setup.ts
