+++ code/core/src/common/versions.ts
@@ -15,10 +15,10 @@ export default {
   '@storybook/addon-onboarding': '8.3.0-beta.0',
   '@storybook/addon-outline': '8.3.0-beta.0',
   '@storybook/addon-storysource': '8.3.0-beta.0',
+  '@storybook/experimental-addon-test': '8.3.0-beta.0',
   '@storybook/addon-themes': '8.3.0-beta.0',
   '@storybook/addon-toolbars': '8.3.0-beta.0',
   '@storybook/addon-viewport': '8.3.0-beta.0',
-  '@storybook/experimental-addon-vitest': '8.3.0-beta.0',
   '@storybook/builder-vite': '8.3.0-beta.0',
   '@storybook/builder-webpack5': '8.3.0-beta.0',
   '@storybook/core': '8.3.0-beta.0',
